pyproject.toml
[tool.poetry]
name = "advent-of-code"
version = "0.1.0"
description = ""
authors = ["kvwu <kvwu@transienterror.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "advent_of_code"}]

[tool.poetry.dependencies]
python = "^3.10"
icecream = "^2.1.3"
pyfunctional = "^1.4.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-benchmark = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
